Separate the snes specifics from the general utility
Places `readAsArrayBuffer` in the root and the snes addressing in an snes module.
